pyproject.toml
[project]
name = "deep-research"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.124.0",
    "greenlet>=3.4.0",
    "mcp>=1.27.0",
    "mcp-server-fetch>=2025.4.7",
    "openai>=2.32.0",
    "openai-agents>=0.14.5",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.38.0",
]
